Refactor error logging in TMSMeters
The 'success' parameter in recordServerRequest and recordClientRequest has been replaced with 'error' to enable capturing of specific exceptions on request failures for more accurate log messages. The recording method has also been refactored to include up to three levels of cause for an exception. The 'success' field in the RequestMeter interface has been replaced with 'result', which will hold either the exception or a successful Unit. With these changes, error logging in TMSMeters becomes more informative and useful for debugging.