Mark HttpClients.create() as deprecated due to thread leaks
A deprecation notice has been added to the HttpClients.create() function due to an observed issue where new clients appear to cause thread leaks even when properly closed. This problem is currently under investigation. An intentional delay has also been injected after client creation in an attempt to expose more potential leaks. This is a temporary measure and will be reviewed once the investigation is complete.