Improve JWT serialization and base64JSON visibility
Added a serializer to JWT to better manage string transformations during serialization. This change provides flexibility in the serialization and deserialization of JWT implementation. Additionally, the visibility of unbase64 function from private to public was changed to support flexible use for other potential operations. Additionally, inline function for JWTManager was added for easy instantiation with default Json and signer parameters.