Refactored Bytes package, revised number conversion and added new file
The Bytes package was refactored to remove unnecessary imports and classes for better efficiency and readability. The unnecessary Bytes4 class and associated functions were removed from Bytes.kt file while the core functionality was preserved and moved to a new separate file, Bytes4.kt, to maintain modularity and improve option for further expansion.

The UInt24.kt was updated to include a new annotation class MoreNumberHelper and a detailed documentation of the UInt24 class was added for better understanding. Changes to the numberConversion.kt file included comment additions to provide more information on the types of number conversions available.

Another major update was the addition of new file, asNumbers.kt, which provides function declarations to convert ByteArray into multiple data types with both big-endian and native byte order options.